Update README.md
Dropping a grapefruit

Suppose Hector drops a grapefruit off a high bridge and lets it fall until it hits the water. If we neglect air resistance, then the distance d in meters fallen by the grapefruit after t seconds is d = 0.5 x g x t^2 , where the acceleration of gravity is the constant g = 9.8 (m/s^2).

Write a program that asks the user to input the number of seconds that a grapefruit falls from a bridge above the water. The program should then calculate the distance fallen for each second from t = 0 until the value of t input by the user and display the output on the console.

Sample Run 1

Please input the time of fall in seconds: 
2

Time Falling (seconds) Distance Fallen (meters)
********************** ************************
0                                     0
1                                     4.9
2                                    19.6

Sample Run 2

Please input the time of fall in seconds: 
4

Time Falling (seconds) Distance Fallen (meters)
********************** ************************
0                                      0
1                                       4.9
2                                      19.6
3                                      44.1
4                                      78.4